Moved the MLTestResultsWindow to a new file
